fix(refactor) individual components
