+++ packages/light-apps/src/Onboarding/Welcome/index.tsx
@@ -2,8 +2,7 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-export * from './CreateNewAccountScreen';
-export * from './ImportOptionsScreen';
-export * from './Onboarding';
-export * from './SaveScreen';
-export * from './Welcome';
+export * from './Intro';
+export * from './ChooseCreationOption';
+export * from './SetupNominator';
+export * from './WelcomeScreen';
